requirements.txt
aioredis==1.3.1
argon2-cffi==19.2.0
asgiref==3.2.3
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==3.0.0
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==3.0.3
django-multiurl==1.4.0
djangorestframework==3.11.0
entrypoints==0.3
Faker==4.0.2
flake8==3.7.9
hiredis==1.0.1
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.0
Pillow==7.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
PyHamcrest==2.0.2
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2019.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
Twisted==20.3.0
txaio==20.4.1
wrapt==1.11.2
zope.interface==5.1.0
